Use resistive touch panel instead of PixyCam
A resistive touch panel measures the position of the steel ball rather
than PixyCam detecting a coloured ball. The steel ball must be ~80g or
heavier to be detected by the touch panel.

I would like to use a lighter ball when tuning the PID controller so
that there is less momentum when the ball falls off the plate :)
Observed that with lighter balls, some readings are dropped from the
touch panel. Rather than using a low-pass filter, the best option seems
to be to  use the last reading when this occurs.

I'm concerned that the servo movement is a bit jerky when running the
plate rotate program. I assume that this means there may be some error
in the servo position. Hopefully the PID algorithm will be able to
account for this (and a non-level starting point) by using a non-zero
KI constant.

This change includes:

- CAD for 3D printing of plate support. Holds touch panel in place and
  attaches to rose joints from servo arms.
- CAD for positioning of servo brackets on wooden base, and cutting
  channels for servo arms to move through.
- Updated angle calculations based on new servo positions, i.e. 0 is at
  the 9 o'clock position, 1 is at the 1 o'clock position and 2 is at the
  5 o'clock position.