Tweaked PID constants and added bluetooth control
Improved responsiveness by doubling KP, and slightly increasing KD.
Higher values of KI or KD result in increased oscillations close to
steady state. The ball is never perfectly still, so the plate is never
still either. The best controller constants have lead to minimal
oscillations, but never eliminated. Deacreasing KD too much results in
instability because the plate is not tilted to brake the speed as the
ball approaches the setpoint.

Also restored the code for bluetooth control and tweaked the setpoints
for the square and circle patterns.