Tweaked position filtering code
The position filtering code would sometimes detect a spike but then the
raw position would have changed rapidly so that it though the readings
were perpetual spikes. Changed the code to ignore a only a single
reading for a spike.

Also moved position filtering code to its own class and experimented
with adding a low pass filter on the ball position output. This gave no
real improvement, but increased the impact of the spike due to settling
time of the filter.

Sometimes we get a double-reading spike, so maybe we can change the code
to allow for that rare occurrence when using the smallest ball.

Also experimented with various PID controller constant values and found
no real improvements over the previous values. Changed the set point
to better reflect the physical centre of the touch screen.

It's highly likely that the imitation MG996R servos are not accurate
enough to perfectly balance the ball. I may try Hitec HS-5485HB servos
instead, as I saw a very impressive ball and plate system on YouTube
using those servos.