Switch to Servo library using new board for power
My new Servo Power PCB arrived and this update switches from the PCA9685
external PWM driver to the Arduino Servo library. This theoretically
allows for higher precision due to the writeMicroseconds function
allowing ~10bit precision instead of ~8bit. However, the servos may not
be that accurate due to deadband.

Also replaced the generic Arduino Nano v3 with a genuine Arduino Nano
Every. The new microcontroller runs at 20MHz, ideally giving a more
accurate PWM signal for the servos, and includes a second serial port
removing the need for the SoftwareSerial library for Bluetooth
communication (although this is still TODO).

Re-tuned the PID controller and allowed a higher maximum angle to allow
KP to be increased further. Using the Servo library seems to have
provided more accuracy given that KD could be doubled. However, the
servos still seem to have a large deadband that leads to the ball never
being stationary, and the servos appear to twich between two positions,
never able to stop in the middle.

The next step is to replace the existing generic MG996R servos with JX
PDI-6113MG servos. The new servos use a coreless motor and promise a 1us
deadband (over the alleged 4us of the MG996R). However, the new servos
are 5mm shorter above the mounting holes, so I may need to move the
mounting brackets.

If the brackets are moved, I will likely also change to a shorter servo
arm (20mm rather than 35mm). This should give more fine grained control
as the ratio between the servo angle and the plate angle will increase.