Adjustments for new Servos
Swapped the generic MG996R for JX PDI-6113MG and made corresponding
changes to the code. The new servos seem to have 90 degrees of rotation
compared to the 180 degrees of the MG996Rs. Changed the formula to
compute the angle accordingly, and changed the centre positions for the
new servos.

The end result was a bit of a disaster. The new servos are very jittery
(and noisy), and do not hold the plate position well. With the power
off, the plate slowly sinks until it reaches some kind of equilibrium.
With the power on, the servos and the plate vibrate a lot. When
controlling the ball, the vibrations sometimes get more violent,
especially when the ball nears the edges of the plate. The ball
sometimes bounces due to these vibrations, but luckily there seems to
be no damage to the plate.

The larger vibrations seem like the servos might be exceeding their
stall torque, but that seems unlikely since each of the 3 is rated for
11 kg-cm, which is not much less than the 15 kg-cm of the MG996R. Maybe
the JX PDI-6113MG is great in an RC car, but it doesn't seem suited to
this application.

Next step is to go back to the MG996R, but use a shorter servo arm to
try to make the control more accurate.