Updated constants for short servo arms
Switched back to MG996R servos with short servo arms (20mm instead of
37mm). The end result is that the ball stays closer to the setpoint
than before. There is still some continual movement of the ball, but the
plate does not move as much as with the longer servo arms. The effect
error in the servo position has been reduced as intended. It's not
perfect, but it's an improvement.

I will try to compute the formula for how the servo angle affects the
plate angle. If this looks non-linear, then I may reverse this so that
the PID controller sets the plate angle, which is then translated into
the servo angle. This may improve the performance of the PID controller
by making the system seem more linear.

I may try the JX PDI-6113MG servos again with the short servo arms to
see if they perform better, or I may call it done and start work on the
Android app to control the setpoint via bluetooth.