Replaced servo[1] with spare MG996R
Servo[1] was exhibiting some jerkiness when moving and seemed to be less
precise than the others. I replaced it with the spare MG996R and the
result was good. The ball stays closer to the setpoint than before.

I then gave the JX servos one last try, thinking that the reduced torque
due to the shorter servo arms may reduce the jitter. However, even with
no ball on the plate, the servos click like a geiger counter. Placing
the ball on the plate makes the servos louder, and they still jump about
when the ball is close to the edge of the plate. This happens less than
with the longer servo arms, but the result is still unacceptable.

I then reverted back to the MG996R servos. The ball seems to move around
the set point more than I remember, but I can't be 100% sure. I will try
further tuning of the PID controller, but I'm happy overall with the
result. The only way to do better would be to use Dynamixel servos which
are designed for accuracy.